requirements.txt
asgiref==3.5.2
beautifulsoup4==4.12.2
certifi==2022.9.24
charset-normalizer==3.1.0
distlib==0.3.6
Django==4.1.1
filelock==3.8.0
idna==3.4
oauthlib==3.2.2
Pillow==9.5.0
pipenv==2022.9.24
platformdirs==2.5.2
requests==2.30.0
requests-oauthlib==1.3.1
schedule==1.2.0
soupsieve==2.4.1
sqlparse==0.4.3
tweepy==4.14.0
tzdata==2022.4
urllib3==2.0.2
virtualenv==20.16.5
virtualenv-clone==0.5.7
virtualenvwrapper-win==1.2.7


# note not all packages are required, the project was installed globally on my Machine

# required packages includes
# 1. beautifulsoup4
# 2. requests
# 3. schedule
# 4. tweepy
